pyproject.toml
[tool.poetry]
name = "techfriendaibackend"
version = "0.1.0"
description = ""
authors = ["actusnileh <actusnileh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.114.0"
uvicorn = "^0.30.6"
pydantic = "^2.9.0"
pydantic-settings = "^2.4.0"
sqlalchemy = "^2.0.35"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
httpx = "^0.27.2"
numpy = "^2.1.1"
torch = "^2.4.1"
transformers = "^4.44.2"
celery = "^5.4.0"
pillow = "^10.4.0"
redis = "^5.0.8"
sentencepiece = "^0.2.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output=3
include_trailing_comma=true
line_length=120
lines_after_imports=2
lines_between_sections=1
lines_between_types=0
force_alphabetical_sort_within_sections=true
ensure_newline_before_comments=true
combine_as_imports=true
combine_star=true
skip_gitignore=true
balanced_wrapping=true
force_grid_wrap=2
src_paths=["app", "tests"]
skip_glob=[
    ".git",
    ".idea",
    ".vscode",
    "venv",
    ".venv",
]
known_fastapi=["fastapi"]
known_first_party=["app"]
sections=[
    "FUTURE",
    "STDLIB",
    "FASTAPI",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]